# Optimize GetSignData