--- a/src/Neo/UInt256.cs
+++ b/src/Neo/UInt256.cs
@@ -12,6 +12,7 @@
 using Neo.Extensions;
 using Neo.IO;
 using System;
+using System.Buffers.Binary;
 using System.IO;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
@@ -128,6 +129,21 @@ namespace Neo
             writer.Write(value2);
             writer.Write(value3);
             writer.Write(value4);
+        }
+
+        public void Serialize(Span<byte> destination)
+        {
+            const int IxValue2 = sizeof(ulong);
+            const int IxValue3 = sizeof(ulong) * 2;
+            const int IxValue4 = sizeof(ulong) * 3;
+
+            Span<byte> buffer = stackalloc byte[sizeof(ulong) * 4];
+            BinaryPrimitives.WriteUInt64LittleEndian(buffer, value1);
+            BinaryPrimitives.WriteUInt64LittleEndian(buffer[IxValue2..], value2);
+            BinaryPrimitives.WriteUInt64LittleEndian(buffer[IxValue3..], value3);
+            BinaryPrimitives.WriteUInt64LittleEndian(buffer[IxValue4..], value4);
+
+            buffer.CopyTo(destination);
         }
 
         public override string ToString()