--- a/tests/Neo.UnitTests/UT_Helper.cs
+++ b/tests/Neo.UnitTests/UT_Helper.cs
@@ -30,7 +30,7 @@ namespace Neo.UnitTests
         public void GetSignData()
         {
             TestVerifiable verifiable = new();
-            byte[] res = verifiable.GetSignData(TestProtocolSettings.Default.Network);
+            var res = verifiable.GetSignData(TestProtocolSettings.Default.Network);
             Assert.AreEqual("4e454f3350b51da6bb366be3ea50140cda45ba7df575287c0371000b2037ed3898ff8bf5", res.ToHexString());
         }