@shargon 2025-03-04
With the last change and following the logic of `GetSpan` it improve 10 times

```
| Method  | Mean      | Error     | StdDev    |
|-------- |----------:|----------:|----------:|
| TestOld | 76.315 ns | 1.1920 ns | 1.0566 ns |
| TestNew |  7.081 ns | 0.1683 ns | 0.1492 ns |
```